requirements.txt
llama-index
langchain
langchain-community
langchain_huggingface
PyPDF2
faiss-cpu
sentence-transformers
openai
transformers
accelerate
bitsandbytes
streamlit
python-dotenv